Removes the NonInteractableTag.
