+++ Controls/Systems/Messaging/ButtonCloseSystem.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5de8699a41b85d04f957bcf436ed6e09
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
