+++ Controls/Systems/Messaging/ButtonMessageSystem.cs
@@ -1,33 +1,6 @@
-﻿using UGUIDots.Render;
-using Unity.Entities;
+﻿using Unity.Entities;
 
 namespace UGUIDots.Controls.Messaging.Systems {
-
-    [UpdateInGroup(typeof(MessagingUpdateGroup))]
-    public class ButtonCloseSystem : SystemBase {
-
-        private EntityCommandBufferSystem cmdBufferSystem;
-
-        protected override void OnCreate() {
-            cmdBufferSystem = World.GetOrCreateSystem<BeginPresentationEntityCommandBufferSystem>();
-        }
-
-        protected override void OnUpdate() {
-            var cmdBuffer = cmdBufferSystem.CreateCommandBuffer();
-            var disabled = GetComponentDataFromEntity<Disabled>();
-            Entities.WithAll<ButtonMessageRequest>().ForEach((in CloseTarget c0) => {
-                var targetEntity = c0.Value;
-
-                if (disabled.Exists(targetEntity)) {
-                    cmdBuffer.RemoveComponent<Disabled>(targetEntity);
-                    cmdBuffer.AddComponent<EnableRenderingTag>(targetEntity);
-                } else {
-                    cmdBuffer.AddComponent(targetEntity, new Disabled { });
-                    cmdBuffer.AddComponent<DisableRenderingTag>(targetEntity);
-                }
-            }).Run();
-        }
-    }
 
     // TODO: Rename the accompanying file
     [UpdateInGroup(typeof(MessagingConsumptionGroup))]
